Midi channels are now ints
But when setting them, the value is always
checked to be in range [0..15].

+++ Moritz.Algorithm/Song Six/Clytemnestra.cs
@@ -9,7 +9,7 @@ namespace Moritz.Algorithm.SongSix
 	/// </summary>
 	internal class Clytemnestra : Trk
     {
-        internal Clytemnestra(byte midiChannel, Trk wind3) 
+        internal Clytemnestra(int midiChannel, Trk wind3) 
             : base(midiChannel, new List<IUniqueDef>())
         {
             SetMomentDefsListPerVerse();
